% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pe.R
\name{test_pe}
\alias{test_pe}
\title{Calculate PE of a time series and its statistical significance.}
\usage{
test_pe(x, m = 3, tau = 1, n_random = 100)
}
\arguments{
\item{x}{A numeric vector.}

\item{m}{Word length.}

\item{tau}{Time lag.}

\item{n_random}{Number of random permutations of the time series to use in randomzied permutation test.}
}
\value{
a numeric vector with PE, number of permutations (NP) used to calculate PE, and the statistical significance of PE.
}
\description{
Calculate PE of a time series and its statistical significance against the null hypothesis that the target time series is not more predictable than white noise.
}
\examples{
x <- rnorm(50)
test_pe(x)
}
